# src/main/gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: API specification for authentication and user management endpoints.
  title: Identity Service API
  version: 1.0.0
servers:
- description: Local development
  url: http://localhost:8080
paths:
  /info:
    get:
      description: Returns a short string indicating the service is running.
      responses:
        "200":
          content:
            text/plain:
              schema:
                example: Service is working.
                type: string
          description: Service info string.
      summary: Health/info endpoint
      tags:
      - info
      x-accepts:
      - text/plain
  /auth/signup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRegister'
        required: true
      responses:
        "201":
          description: User created.
        "400":
          description: Validation error.
      summary: Register a new user
      tags:
      - auth
      x-content-type: application/json
      x-accepts:
      - application/json
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
        required: true
      responses:
        "200":
          description: Login accepted.
        "401":
          description: Invalid credentials.
      summary: Authenticate a user
      tags:
      - auth
      x-content-type: application/json
      x-accepts:
      - application/json
  /auth/reset:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
      responses:
        "200":
          description: Password reset initiated.
        "400":
          description: Validation error.
      summary: Reset a user's password
      tags:
      - auth
      x-content-type: application/json
      x-accepts:
      - application/json
  /user/addUser:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRegister'
        required: true
      responses:
        "201":
          description: User created.
        "400":
          description: Validation error.
      summary: Create a user (admin)
      tags:
      - users
      x-content-type: application/json
      x-accepts:
      - application/json
  /user/deleteUser/{username}:
    post:
      parameters:
      - explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: User deleted.
        "404":
          description: User not found.
      summary: Delete a user by username
      tags:
      - users
      x-accepts:
      - application/json
  /user/deleteAllUsers:
    post:
      responses:
        "200":
          description: All users deleted.
      summary: Delete all users
      tags:
      - users
      x-accepts:
      - application/json
  /user/getAllUsers:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
          description: List of users.
      summary: List all users
      tags:
      - users
      x-accepts:
      - application/json
components:
  schemas:
    SignUpRegister:
      example:
        password: P@ssw0rd!
        email: john.doe@example.com
        username: johndoe
      properties:
        username:
          example: johndoe
          type: string
        email:
          example: john.doe@example.com
          format: email
          type: string
        password:
          example: P@ssw0rd!
          format: password
          type: string
      required:
      - email
      - password
      - username
      type: object
    SignInRequest:
      example:
        password: P@ssw0rd!
        login: johndoe
      properties:
        login:
          description: Username or email.
          example: johndoe
          type: string
        password:
          example: P@ssw0rd!
          format: password
          type: string
      required:
      - login
      - password
      type: object
    ResetPasswordRequest:
      example:
        password: NewP@ssw0rd!
        login: johndoe
      properties:
        login:
          description: Username or email.
          example: johndoe
          type: string
        password:
          example: NewP@ssw0rd!
          format: password
          type: string
      required:
      - login
      - password
      type: object
    UserResponse:
      example:
        id: 1
        email: john.doe@example.com
        username: johndoe
      properties:
        id:
          example: 1
          format: int64
          type: integer
        email:
          example: john.doe@example.com
          format: email
          type: string
        username:
          example: johndoe
          type: string
      type: object